# types: make chopPrecisionAndTruncate always non-mutative